Mega plus cleanliness (#8)
* Revert d3086855486749b6f9d3d569e00c05b52eab2ffd

This is a dead file ramaining from a rebase, after the fix was
merged upstream with some changes.

* Revert 89675dd81a6e59ef63e458b0258304eb31264cc6

* Fix tab spacing

* USB-Audio: Add Profile for ASUS ROG Strix Z590-A Gaming WiFi

Asus released board(s) that use the same ALC4080, but have slightly
different mappings for mic & line-in, as well they don't have a
SPDIF connector.

* USB-Audio: Add quirk for Asus Pro WRX80E SAGE SE WIFI

Most of the audio jacks work out of the box, the line-in and
mic-in need to be remapped though.

* Hack: Mira B1 mic/line-in to work

This is because of me trying to simplify the code in a previous
commit. This isn't perfect, but new features in alsa-ucm-conf's
config files are coming in the next version, so this change will
have to be re-written anyway. Thus, this is good enough for now.